+++ tests/metrics_publishing.rs
@@ -4,9 +4,8 @@ use opentelemetry_sdk::{
     metrics::{
         data::{self, Gauge, Histogram, Sum},
         reader::MetricReader,
-        InstrumentKind, ManualReader, MeterProviderBuilder, MetricError, SdkMeterProvider,
+        InstrumentKind, ManualReader, MeterProviderBuilder, SdkMeterProvider,
     },
-    Resource,
 };
 
 use std::{fmt::Debug, sync::Arc};
@@ -548,10 +547,7 @@ impl MetricReader for TestReader {
         self.inner.register_pipeline(pipeline);
     }
 
-    fn collect(
-        &self,
-        rm: &mut data::ResourceMetrics,
-    ) -> opentelemetry_sdk::metrics::MetricResult<()> {
+    fn collect(&self, rm: &mut data::ResourceMetrics) -> OTelSdkResult {
         self.inner.collect(rm)
     }
 
@@ -566,6 +562,10 @@ impl MetricReader for TestReader {
     fn temporality(&self, kind: InstrumentKind) -> opentelemetry_sdk::metrics::Temporality {
         self.inner.temporality(kind)
     }
+
+    fn shutdown_with_timeout(&self, timeout: std::time::Duration) -> OTelSdkResult {
+        self.inner.shutdown_with_timeout(timeout)
+    }
 }
 
 struct TestExporter<T> {
@@ -577,78 +577,105 @@ struct TestExporter<T> {
     _meter_provider: SdkMeterProvider,
 }
 
+trait AsAny {
+    fn as_any(&self) -> &dyn std::any::Any;
+}
+
+impl AsAny for data::AggregatedMetrics {
+    fn as_any(&self) -> &dyn std::any::Any {
+        match self {
+            data::AggregatedMetrics::F64(x) => match x {
+                data::MetricData::Gauge(x) => x as &dyn std::any::Any,
+                data::MetricData::Sum(x) => x as &dyn std::any::Any,
+                data::MetricData::Histogram(x) => x as &dyn std::any::Any,
+                data::MetricData::ExponentialHistogram(x) => x as &dyn std::any::Any,
+            },
+            data::AggregatedMetrics::U64(x) => match x {
+                data::MetricData::Gauge(x) => x as &dyn std::any::Any,
+                data::MetricData::Sum(x) => x as &dyn std::any::Any,
+                data::MetricData::Histogram(x) => x as &dyn std::any::Any,
+                data::MetricData::ExponentialHistogram(x) => x as &dyn std::any::Any,
+            },
+            data::AggregatedMetrics::I64(x) => match x {
+                data::MetricData::Gauge(x) => x as &dyn std::any::Any,
+                data::MetricData::Sum(x) => x as &dyn std::any::Any,
+                data::MetricData::Histogram(x) => x as &dyn std::any::Any,
+                data::MetricData::ExponentialHistogram(x) => x as &dyn std::any::Any,
+            },
+        }
+    }
+}
+
 impl<T> TestExporter<T>
 where
     T: Debug + PartialEq + Copy + std::iter::Sum + 'static,
 {
-    fn export(&self) -> Result<(), MetricError> {
-        let mut rm = data::ResourceMetrics {
-            resource: Resource::builder().build(),
-            scope_metrics: Vec::new(),
-        };
+    fn export(&self) -> OTelSdkResult {
+        let mut rm = data::ResourceMetrics::default();
         self.reader.collect(&mut rm)?;
 
-        assert!(!rm.scope_metrics.is_empty());
+        let mut scope_metrics = rm.scope_metrics().peekable();
 
-        rm.scope_metrics.into_iter().for_each(|scope_metrics| {
-            assert_eq!(scope_metrics.scope.name(), INSTRUMENTATION_LIBRARY_NAME);
-            assert_eq!(scope_metrics.scope.version().unwrap(), CARGO_PKG_VERSION);
+        assert!(scope_metrics.peek().is_some());
 
-            scope_metrics.metrics.into_iter().for_each(|metric| {
-                assert_eq!(metric.name, self.expected_metric_name);
+        scope_metrics.for_each(|scope_metrics| {
+            assert_eq!(scope_metrics.scope().name(), INSTRUMENTATION_LIBRARY_NAME);
+            assert_eq!(scope_metrics.scope().version().unwrap(), CARGO_PKG_VERSION);
+
+            scope_metrics.metrics().for_each(|metric| {
+                assert_eq!(metric.name(), self.expected_metric_name);
 
                 match self.expected_instrument_kind {
                     InstrumentKind::Counter | InstrumentKind::UpDownCounter => {
-                        let sum = metric.data.as_any().downcast_ref::<Sum<T>>().unwrap();
+                        let sum = metric.data().as_any().downcast_ref::<Sum<T>>().unwrap();
                         assert_eq!(
                             self.expected_value,
-                            sum.data_points
-                                .iter()
-                                .map(|data_point| data_point.value)
-                                .sum()
+                            sum.data_points().map(|data_point| data_point.value()).sum()
                         );
 
                         if let Some(expected_attributes) = self.expected_attributes.as_ref() {
-                            sum.data_points.iter().for_each(|data_point| {
+                            sum.data_points().for_each(|data_point| {
                                 assert!(compare_attributes(
                                     expected_attributes,
-                                    &data_point.attributes,
+                                    data_point.attributes().cloned().collect(),
                                 ))
                             });
                         }
                     }
                     InstrumentKind::Gauge => {
-                        let gauge = metric.data.as_any().downcast_ref::<Gauge<T>>().unwrap();
+                        let gauge = metric.data().as_any().downcast_ref::<Gauge<T>>().unwrap();
                         assert_eq!(
                             self.expected_value,
                             gauge
-                                .data_points
-                                .iter()
-                                .map(|data_point| data_point.value)
-                                .next_back()
+                                .data_points()
+                                .map(|data_point| data_point.value())
+                                .last()
                                 .unwrap()
                         );
 
                         if let Some(expected_attributes) = self.expected_attributes.as_ref() {
-                            gauge.data_points.iter().for_each(|data_point| {
+                            gauge.data_points().for_each(|data_point| {
                                 assert!(compare_attributes(
                                     expected_attributes,
-                                    &data_point.attributes,
+                                    data_point.attributes().cloned().collect(),
                                 ))
                             });
                         }
                     }
                     InstrumentKind::Histogram => {
-                        let histogram =
-                            metric.data.as_any().downcast_ref::<Histogram<T>>().unwrap();
-                        let histogram_data = histogram.data_points.first().unwrap();
-                        assert!(histogram_data.count > 0);
-                        assert_eq!(histogram_data.sum, self.expected_value);
+                        let histogram = metric
+                            .data()
+                            .as_any()
+                            .downcast_ref::<Histogram<T>>()
+                            .unwrap();
+                        let histogram_data = histogram.data_points().next().unwrap();
+                        assert!(histogram_data.count() > 0);
+                        assert_eq!(histogram_data.sum(), self.expected_value);
 
                         if let Some(expected_attributes) = self.expected_attributes.as_ref() {
                             assert!(compare_attributes(
                                 expected_attributes,
-                                &histogram_data.attributes
+                                histogram_data.attributes().cloned().collect(),
                             ))
                         }
                     }
@@ -666,7 +693,7 @@ where
 // After sorting the KeyValue vec, compare them.
 // Return true if they are equal.
 #[allow(clippy::ptr_arg)]
-fn compare_attributes(expected: &Vec<KeyValue>, actual: &Vec<KeyValue>) -> bool {
+fn compare_attributes(expected: &Vec<KeyValue>, actual: Vec<KeyValue>) -> bool {
     let mut expected = expected.clone();
     let mut actual = actual.clone();